java/src/main/java/array_and_string/Spiral _Matrix/SolutionTest.java
package array_and_string.Spiral_Matrix;

import org.junit.After;
import org.junit.Before;
import org.junit.Rule;
import org.junit.Test;
import org.junit.rules.Timeout;

import java.util.ArrayList;
import java.util.Arrays;

import static org.junit.Assert.assertArrayEquals;
import static org.junit.Assert.assertEquals;

public class SolutionTest {
    @Rule
    public Timeout globalTimeout = new Timeout(200);
    private array_and_string.Spiral_Matrix.Solution solution;

    @Before
    public void setUp() throws Exception {
        solution = new array_and_string.Spiral_Matrix.Solution();
    }

    @After
    public void tearDown() throws Exception {
        solution = null;
    }

    @Test
    public void Test0() {
        int[][] nums = {{1, 2, 3, 4}, {5, 6, 7, 8}, {9, 10, 11, 12}};
        var target = Arrays.asList(1, 2, 3, 4, 8, 12, 11, 10, 9, 5, 6, 7);
        assertEquals(target, solution.spiralOrder(nums));
    }


}



